Atualizações

+++ UFV/aula05/ex07.py
@@ -1,7 +1,7 @@
 nome = str(input('Atleta: ')).strip().title()
 notas = list()
 for contador in range(0, 7):
-    notas.append(int(input(f'Nota {contador + 1}: ')))
+    notas.append(float(input(f'Nota {contador + 1}: ')))
 soma_notas = sum(notas)
 soma_notas -= max(notas) + min(notas)
 media = soma_notas / 7
